+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/extensions/ChatClient.kt
@@ -10,6 +10,7 @@ import io.getstream.chat.android.client.api.models.QuerySort
 import io.getstream.chat.android.client.api.models.SendActionRequest
 import io.getstream.chat.android.client.call.Call
 import io.getstream.chat.android.client.call.CoroutineCall
+import io.getstream.chat.android.client.call.await
 import io.getstream.chat.android.client.events.ChatEvent
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.Channel
@@ -18,6 +19,7 @@ import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.Result
 import io.getstream.chat.android.client.utils.SyncStatus
+import io.getstream.chat.android.client.utils.toUnitResult
 import io.getstream.chat.android.offline.ChatDomain
 import io.getstream.chat.android.offline.ChatDomainImpl
 import io.getstream.chat.android.offline.channel.CreateChannelService
@@ -302,4 +304,16 @@ internal fun ChatClient.shuffleGiphy(message: Message): Call<Message> {
             Result(result.error())
         }
     }
+}
+
+/**
+ * Deletes the channel with the specified id.
+ *
+ * @param cid The full channel id i. e. messaging:123.
+ *
+ * @return Executable async [Call] deleting a channel.
+ */
+@CheckResult
+public fun ChatClient.deleteChannel(cid: String): Call<Unit> = CoroutineCall(domainImpl().scope) {
+    channel(cid).delete().await().toUnitResult()
 }